Refactor: CardForm 로직 분리
