+++ src/components/Button/Button.test.tsx
@@ -2,7 +2,7 @@ import { screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { composeStories } from '@storybook/testing-react';
 //
-import TestA11y from '@/libs/accessibility';
+import TestA11y from '@/utils/accessibility';
 import * as ButtonStories from './Button.stories';
 
 const { DefaultTemplate } = composeStories(ButtonStories);
