+++ src/components/Form/Form.tsx
@@ -1,8 +1,5 @@
+import { FormGetNameValidatorCallbackProps } from 'components';
 import { ChangeEvent, FormEvent, ReactNode, FocusEvent, forwardRef } from 'react';
-import type {
-  FormSameNameFromTargetValidatorProps,
-  FormSameNameFromTargetValidatorCallbackProps,
-} from 'components';
 
 type FormProps = {
   action?: string;
@@ -75,34 +72,54 @@ const generateElementValue = (formElement: HTMLFormElement) => {
   }, {} as Record<string, string[]>);
 };
 
+const nextInputFocus = (event: FocusEvent<HTMLFormElement> | ChangeEvent<HTMLFormElement>) => {
+  const $elements = event.currentTarget.elements;
+  return (index: number) => {
+    if ($elements.length <= index) return;
+
+    const $nextElement = $elements[index + 1] as HTMLElement;
+    $nextElement.focus();
+  };
+};
+
 const isInputElement = (
   $target: (EventTarget & HTMLFormElement) | HTMLInputElement,
 ): $target is HTMLInputElement => $target instanceof HTMLInputElement;
 
-const sameNameFromTargetValidator = (
-  { $elements, $target, name }: FormSameNameFromTargetValidatorProps,
-  callback: (props: FormSameNameFromTargetValidatorCallbackProps) => void,
+const getNameValidator = (
+  event: FocusEvent<HTMLFormElement> | ChangeEvent<HTMLFormElement>,
+  name: string,
+  callback: (data: FormGetNameValidatorCallbackProps) => void,
 ) => {
-  if ($target.name !== name) return;
+  try {
+    const $target = event.target;
+    if (!isInputElement($target)) return;
+    if ($target.name !== name) return;
 
-  const targetIndex = [...$elements].findIndex((element) => element === $target);
+    const $elements = event.currentTarget.elements;
+    const index = [...$elements].findIndex((element) => element === $target);
 
-  const sameNamesElements = [...$elements].filter((element) =>
-    (element as HTMLInputElement).name.includes($target.name),
-  ) as HTMLInputElement[];
+    const sameNames = [...$elements].filter((element) =>
+      (element as HTMLInputElement).name.includes($target.name),
+    ) as HTMLInputElement[];
 
-  if (targetIndex === -1) return;
+    if (index === -1) return;
 
-  callback && callback({ $elements, $target, name, sameNamesElements, targetIndex });
+    callback && callback({ target: $target, sameNames, index });
+  } catch (error) {
+    console.log(error);
+  }
 };
 
 type CompoundedType = typeof Form & {
   isInputElement: typeof isInputElement;
-  sameNameFromTargetValidator: typeof sameNameFromTargetValidator;
+  getNameValidator: typeof getNameValidator;
+  nextInputFocus: typeof nextInputFocus;
 };
 
 const CompoundedForm = Form as CompoundedType;
 CompoundedForm.isInputElement = isInputElement;
-CompoundedForm.sameNameFromTargetValidator = sameNameFromTargetValidator;
+CompoundedForm.getNameValidator = getNameValidator;
+CompoundedForm.nextInputFocus = nextInputFocus;
 
 export default CompoundedForm;
