--- src/libs/utils.ts
@@ -1,2 +1,0 @@
-export const flip = <T extends string, P extends string>(data: Record<T, P>): Record<P, T> =>
-  Object.fromEntries(Object.entries(data).map(([key, value]) => [value, key]));
