+++ src/templates/CardAddPage/index.ts
@@ -1,1 +1,1 @@
-export {};
+export { default as CardForm } from './CardForm';
