+++ src/templates/HomePage/index.ts
@@ -1,1 +1,1 @@
-export {};
+export { default as CardList } from './CardList';
